Refactor tiles
Since we have so many combinations of tiles, it would be impractical
to have one class for each tile. Instead, the base Tile class will be
empty, and special classes will extend it. Textures and open edges are
passed into the constructor.

The textures will all be loaded in at the top level file.