Initial additions for logging.
Right now, the basis of creating the logging object, and some of the more
straightfoward stuff has been added. I still need to add the provided
file into the project, which i'll need access to a linux machine
to more easily get done tomorrow. Otherwise, next step is to figure
out how to pass the logger around to the specific areas, and how to
grab necessary information for certain areas, and determine
which functions are there. Python functions will be kept separate from the
project since we need those on the database more.