Implement objective tiles
Objective tiles are the general class for the key/door mechanic.
Objectives tiles have two special properties: key and prerequisites. The key
identifies the objective uniquely. Prerequisites are the tiles that need
to be visited before the current objective tile can be visited. They are
stored as an array in each tile.

The state of each objective tile is stored in Floor. The objectiveState
dictionary maps objective tile keys to a boolean representing whether they
have been visited.