Add reset function for each tile
The healing tile and objective tile contain a bit of state - the image
disappears when health is used or the objective is reached, and the
healing tile has a flag indicating if its been used. We need to reset this
stuff to the original state when the whole floor is reset as well.