Added logging files for logging and set up initial logging.
Just added the swc library file and the logger action script
file to the project so that logging should work. Added initial
set up for starting a level, and at least logging when restart
is clicked. More actions will be added, including level finished
once those actions are added to the functionality as well. Until
then, the python scripts will be modified and cleaned up so those
are ready to go to grab specific pieces of data to read.