Implement parts of the AI framework (WIP)
Based on the CSE 473 Pacman AI code. I'm defining the game as a search
problem. It's a graph where nodes represent the whole state of the game,
and the edges represent actions the character can take. This can be used
to implement algorithms like A*.

Since we need to make many copies of game state (grid, characters, etc.),
I created several classes to represent the game state separate from the
stuff we already created. These are essentially structs that have no
functionality, and are not sprites. This makes them a bit easier to
manipulate. Also, I divided Tiles into TileState and EntityState -
TileState just represents the walls, and EntityState represents enemies,
objectives, and healing.

We may want to refactor a lot of our stuff to use the State classes, to
avoid copying functionality between the AI module and the Sprites.