Implement rudimentary combat
The character and enemy will now attack back and forth, trading damage.
Currently for visuals I'm just dumping attacks in the middle of the
screen. If the character wins the enemy disappears and the character can
move again.

Still TODO: proper combat info, integration with HUD, what happens when
the player fails.