Added fog of war
Added the fog of war element to the game. All tiles besides those
within a line of sight of 2 (1 diagonally) are visible, and even
highlight that tile won't show anything. Once a tile is placed, then
parts of the map will open up. Still not sure if we want the exit tile
to remove fog as well, and to be able to add a tile to a fog spot, mainly
the exit issue.