Make Heap SearchNode-specific and add priority function
The Heap will only be used to hold SearchNodes, so change all Object types
to SearchNode so we don't have to cast later. Also build in support for
heuristics into the Heap - nodes will be sorted based on the priority
function. The priority function for A* is the movement cost + heuristic
value.

Also add better documentation for the heap.