Add hash functions to state classes.
We need to be able to hash GameState to store them in a Dictionary. We
want to store visited GameStates, but if we use GameState as a key in the
visited Dictionary, it won't work because it compares key by reference.
Therefore we need to make the key a string or int, by hashing.