Fixed the tileInfo positioning
Fixed tile info positioning so that if the user
pans the grid around, the tile info should stay
in place in the top right hand corner now, and not
move around with the grid. Still issues with
if the user moves, the cursor still thinks that
it is highlighting that tile when it isn't anymore.