Refactor Clickable to remove arbitrary parameter arrays
Clickables that need additional parameters or data now have their own
subclasses. Using arrays as parameter lists is bad because it requires us
to write our own parameter checking, when it could be handled by the
compiler. It is hard to debug and read when the parameters are not strictly
defined in the code. Plus, proper static types is self-documenting, unlike
arrays of parameters.