Fix bugs / poor conventions. See description.
1. BuildHUD now has a state string which is used
to verify actions rather than checking for the
current image that is used
2. Highlights on the map now are properly cleared
upon switching between selection types.
3. BuildHUD now properly sets internal state
on switching to delete tool
4. Delete tool now plays some sound when
removing tiles
5. FLOOR_RESET sfx renamed to TILE_REMOVE
6. Redid UI to group buttons a little better, also
in light of the SHOP button being replaced with
a more informative clickable in the future
7. New shop button
8. BuildHUD internal state string is now
also used in various functions to allow
a) delete to have an associated cost
b) ensure that if functions are erroneously called
that they will not result in unintended functionality
9. Removed poor short-circuit in floor's
buildHUD handling method.