Added monster movement
Added monster movement to the game, allowing monsters to
move around the dungeon as well. Works with character movement,
and randomly picks a direction that isn't just backwards when it can.

Enemy now has an optional boolean parameter that sets whether the
monster can move or not. Also added some reset functionality to the
Enemy class for reseting their positions to the initial spots.