Refactor Game.as
Split Game.as into several different classes to better isolate responsibility
between classes. What was all packaged in Game is now:
- Main (overall parent node)
- Menu (displays the main menu)
- Credits (displays the credits screen)
- Game (parent node for all gameplay)

Main is the primary class that coordinates all of the rest. Main will swap in
Menu, Credits, and Game Sprites, which are the parents of everything else. This
substantially simplifies the UI logic. The only UI stored in Main are the mute
buttons and the cursor, which are present on every screen and are always on top.

The previous main class has been renamed to Preloader, as fits its purpose.