More updates to scores
Added more updates to scores, making some functions public
so we can save at the end of each build phase and each
run phase now, and to keep track of gold spent. Also
added some code in for displaying current scores on the menu.

This commit is more to pull code for updating and fixing credits button
that will help fix scores button as well.