Fixed bugs associated with reseting floor and saving
Fixed two bugs associated with reseting a floor and saving.
When resetting a floor, if too many enemies were around,
occasionally one or two enemies wouldn't exist on the entity
grid because of the original set up of reseting, and so they
would exist but not at the same time, and removed on the next
time playing.

The saving issue dealt with the fact that we saved
everything into objective, but loaded them through
prereqs instead.